+++ tests/test_objects.py
@@ -2,9 +2,11 @@ import pytest
 
 from bbconf.exceptions import BEDFileNotFoundError, MissingThumbnailError
 
+from .conftest import SERVICE_UNAVAILABLE
 from .utils import BED_TEST_ID, ContextManagerDBTesting
 
 
+@pytest.mark.skipif(SERVICE_UNAVAILABLE, reason="Database is not available")
 class TestObjects:
     def test_object_path(self, bbagent_obj):
         with ContextManagerDBTesting(config=bbagent_obj.config, add_data=True):
@@ -43,4 +45,8 @@ class TestObjects:
 
 @pytest.mark.skip("Used to visualize the schema")
 def test_create_schema_graph(bbagent_obj):
+    f = ContextManagerDBTesting(config=bbagent_obj.config, add_data=True)
+    f._add_data()
+    f._add_universe()
+
     bbagent_obj.config.db_engine.create_schema_graph()